% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_BIRCH.R
\name{BIRCH-initialize}
\alias{BIRCH-initialize}
\title{CF-Tree creation}
\arguments{
\item{treshold}{Specifies the treshold used to check whether a new datapoint can be absorbed or not.}

\item{branching}{Specifies the branching factor (maximum amount of child nodes for a nonleaf node) of the CF-Tree.}

\item{maxLeaf}{Specifies the maximum number of entries within a leaf node.}

\item{maxMemory}{Specifies the memory limitation for the whole CFTree in bytes. Default is 0, indicating no memory restriction.}

\item{outlierThreshold}{Specifies the threshold for identifying outliers when rebuilding the CF-Tree.}
}

