% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{CellQC}
\alias{CellQC}
\title{Spatial single cell quality control}
\usage{
CellQC(object, ...)
}
\arguments{
\item{object}{An Giotto object}

\item{...}{Arguments passed to other methods}
}
\description{
Spatial single cell quality control
}
