% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD.R
\name{close_stream}
\alias{close_stream}
\title{Close a Data Stream}
\usage{
close_stream(dsd, ...)
}
\arguments{
\item{dsd}{An object of class a subclass of \link{DSD} which implements a
reset function.}

\item{...}{further arguments.}
}
\description{
Close a data stream that needs closing (e.g., a file or a connection).
}
\details{
\code{close_stream()} is implemented for:
\itemize{
\item \link{DSD}
\item \link{DSD_ReadCSV}
\item \link{DSD_ReadDB}
\item \link{DSD_ReadStream}
\item \link{DSF}
\item \link{DST_WriteStream}
}
}
\seealso{
Other DSD:
\code{\link[=DSD]{DSD()}},
\code{\link[=DSD_BarsAndGaussians]{DSD_BarsAndGaussians()}},
\code{\link[=DSD_Benchmark]{DSD_Benchmark()}},
\code{\link[=DSD_Cubes]{DSD_Cubes()}},
\code{\link[=DSD_Gaussians]{DSD_Gaussians()}},
\code{\link[=DSD_MG]{DSD_MG()}},
\code{\link[=DSD_Memory]{DSD_Memory()}},
\code{\link[=DSD_Mixture]{DSD_Mixture()}},
\code{\link[=DSD_NULL]{DSD_NULL()}},
\code{\link[=DSD_ReadDB]{DSD_ReadDB()}},
\code{\link[=DSD_ReadStream]{DSD_ReadStream()}},
\code{\link[=DSD_Target]{DSD_Target()}},
\code{\link[=DSD_UniformNoise]{DSD_UniformNoise()}},
\code{\link[=DSD_mlbenchData]{DSD_mlbenchData()}},
\code{\link[=DSD_mlbenchGenerator]{DSD_mlbenchGenerator()}},
\code{\link[=DSF]{DSF()}},
\code{\link[=animate_data]{animate_data()}},
\code{\link[=get_points]{get_points()}},
\code{\link[=plot.DSD]{plot.DSD()}},
\code{\link[=reset_stream]{reset_stream()}}
}
\author{
Michael Hahsler
}
\concept{DSD}
