% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Deconvolution}
\alias{Deconvolution}
\title{Deconvolution of Spatial data}
\usage{
Deconvolution(object, ...)
}
\arguments{
\item{object}{An giotto object}

\item{...}{Argument passed to other methods}
}
\description{
Deconvolution of Spatial data
}
