% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\docType{class}
\name{DefaultEvalCallback-class}
\alias{DefaultEvalCallback-class}
\alias{DefaultEvalCallback}
\title{Default Class for Evaluation Callbacks}
\usage{
DefaultEvalCallback()
}
\description{
The default callback class for the \emph{stream} package. Implicitly
instantiated for each \code{\link{evaluate}} call.
}
\author{
Dalibor Krleža
}
