% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{DiffGenes}
\alias{DiffGenes}
\alias{DiffGenes,SPARK-method}
\alias{DiffGenes,giotto-method}
\title{Different expression with spatial single}
\usage{
DiffGenes(object, ...)

\S4method{DiffGenes}{SPARK}(object, num.cores = 4, verbose = FALSE, ...)

\S4method{DiffGenes}{giotto}(
  object,
  minimum.k = 2,
  maximum.distance.delaunay = 400,
  bin.method = "kmeans",
  expression.value = "normalized",
  num.cores = 5,
  verbose = TRUE,
  python_Path = NULL,
  ...
)
}
\arguments{
\item{object}{A SPARK object}

\item{...}{Other graguments in \code{\link[SPARK]{spark.vc}} 
\code{\link[SPARK]{spark.test}}}

\item{num.cores}{The number of cores to use, default 5.}

\item{verbose}{verbosity of the function}

\item{minimum.k}{minimum nearest neigbhours if maximum_distance != NULL, 
default, 2.}

\item{maximum.distance.delaunay}{distance cutoff for nearest neighbors 
to consider for Delaunay network, default 400}

\item{bin.method}{method to binarize gene expression, choice 
with kmeans and rank, defualt kmeans}

\item{expression.value}{expression values to use, choice with 
normalized, scaled and custom, defualt is normalized.}

\item{python_Path}{python path. Path in Giotto object will be used when no extra dir is provided.}
}
\value{
Different expression data.fame

A list contain theree methods of differentin expression
}
\description{
Different expression with spatial single
}
