% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{DomainPlot}
\alias{DomainPlot}
\alias{DomainPlot,giotto-method}
\title{Visualization of Spatial Clustering}
\usage{
DomainPlot(object, ...)

\S4method{DomainPlot}{giotto}(object, savePlot = TRUE, outputFolder = NULL, ...)
}
\arguments{
\item{object}{a giotto object}

\item{...}{Argument passed to other methods}

\item{savePlot}{Boolean, whether to save plot.}

\item{outputFolder}{Output folder to save results. Path in Giotto object will be used when no external dir is provided.}
}
\description{
Visualization of Spatial Clustering

Spatial domain visualization for cluster result.
}
