% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSClassify.R
\name{DSClassify}
\alias{DSClassify}
\title{Abstract Class for Data Stream Classifiers}
\usage{
DSClassify(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for data stream classifiers. Currently, \pkg{stream} does not
implement classification algorithms.
}
\examples{
DSClassify()
}
\seealso{
\link{DST}
}
\author{
Michael Hahsler
}
