% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_MG.R
\name{DSD_MG}
\alias{DSD_MG}
\alias{add_cluster}
\alias{get_clusters}
\alias{remove_cluster}
\alias{add_cluster.DSD_MG}
\title{DSD Moving Generator}
\usage{
DSD_MG(dimension = 2, ..., labels = NULL, description = NULL)

add_cluster(x, c, label = NULL)

get_clusters(x)

remove_cluster(x, i)

\method{add_cluster}{DSD_MG}(x, c, label = NULL)
}
\arguments{
\item{dimension}{the dimension of the DSD object}

\item{...}{initial set of \link{MGC}s}

\item{description}{An optional string used by \code{print()} to describe the
data generator.}

\item{x}{A \code{DSD_MG} object.}

\item{c}{The cluster that should be added to the \code{DSD_MG} object.}

\item{label, labels}{integer representing the cluster label. \code{NA}
represents noise.  If labels are not specified, then each new cluster gets a
new label.}

\item{i}{The index of the cluster that should be removed from the
\code{DSD_MG} object.}
}
\description{
Creates an evolving DSD consisting of several \link{MGC}s, each representing a moving cluster.
}
\details{
This DSD is able to generate complex datasets that are able to evolve over a
period of time. Its behavior is determined by a set of \link{MGC}s, each representing
a moving cluster.
}
\examples{
### create an empty DSD_MG
stream <- DSD_MG(dimension = 2)
stream

### add two clusters
c1 <- MGC_Random(density = 50, center = c(50, 50), parameter = 1)
add_cluster(stream, c1)
stream

c2 <- MGC_Noise(density = 1, range = rbind(c(-20, 120), c(-20, 120)))
add_cluster(stream, c2)
stream

get_clusters(stream)
get_points(stream, n = 5)
plot(stream, xlim = c(-20,120), ylim = c(-20, 120))

if (interactive()) {
animate_data(stream, n = 5000, xlim = c(-20, 120), ylim = c(-20, 120))
}

### remove cluster 1
remove_cluster(stream, 1)
stream

get_clusters(stream)
plot(stream, xlim = c(-20, 120), ylim = c(-20, 120))

### create a more complicated cluster structure (using 2 clusters with the same
### label to form an L shape)
stream <- DSD_MG(dimension = 2,
  MGC_Static(density = 10, center = c(.5, .2),   parameter = c(.4, .2),
             shape = Shape_Block),
  MGC_Static(density = 10, center = c(.6, .5),   parameter = c(.2, .4),
             shape = Shape_Block),
  MGC_Static(density = 5,  center = c(.39, .53), parameter = c(.16, .35),
             shape = Shape_Block),
  MGC_Noise( density = 1,  range = rbind(c(0,1), c(0,1))),
  labels = c(1, 1, 2, NA)
  )
stream

plot(stream, xlim = c(0, 1), ylim = c(0, 1))

### simulate the clustering of a splitting cluster
c1 <- MGC_Linear(dimension = 2, keyframelist = list(
  keyframe(time = 1,  density = 20, center = c(0,0),   parameter = 10),
  keyframe(time = 50, density = 10, center = c(50,50), parameter = 10),
  keyframe(time = 100,density = 10, center = c(50,100),parameter = 10)
))

### Note: The second cluster appears at time = 50
c2 <- MGC_Linear(dimension = 2, keyframelist = list(
  keyframe(time = 50, density = 10, center = c(50,50), parameter = 10),
  keyframe(time = 100,density = 10, center = c(100,50),parameter = 10)
))

stream <- DSD_MG(dimension = 2, c1, c2)
stream

dbstream <- DSC_DBSTREAM(r = 20, lambda = 0.1)
if (interactive()) {
purity <- animate_cluster(dbstream, stream, n = 2500, type = "micro",
                          xlim = c(-10, 120), ylim = c(-10, 120),
                          measure = "purity", horizon = 100)
}
}
\seealso{
\link{MGC} for types of moving clusters.

Other DSD:
\code{\link[=DSD]{DSD()}},
\code{\link[=DSD_BarsAndGaussians]{DSD_BarsAndGaussians()}},
\code{\link[=DSD_Benchmark]{DSD_Benchmark()}},
\code{\link[=DSD_Cubes]{DSD_Cubes()}},
\code{\link[=DSD_Gaussians]{DSD_Gaussians()}},
\code{\link[=DSD_Memory]{DSD_Memory()}},
\code{\link[=DSD_Mixture]{DSD_Mixture()}},
\code{\link[=DSD_NULL]{DSD_NULL()}},
\code{\link[=DSD_ReadDB]{DSD_ReadDB()}},
\code{\link[=DSD_ReadStream]{DSD_ReadStream()}},
\code{\link[=DSD_Target]{DSD_Target()}},
\code{\link[=DSD_UniformNoise]{DSD_UniformNoise()}},
\code{\link[=DSD_mlbenchData]{DSD_mlbenchData()}},
\code{\link[=DSD_mlbenchGenerator]{DSD_mlbenchGenerator()}},
\code{\link[=DSF]{DSF()}},
\code{\link[=animate_data]{animate_data()}},
\code{\link[=close_stream]{close_stream()}},
\code{\link[=get_points]{get_points()}},
\code{\link[=plot.DSD]{plot.DSD()}},
\code{\link[=reset_stream]{reset_stream()}}
}
\author{
Matthew Bolanos
}
\concept{DSD}
