% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_Mixture.R
\name{DSD_Mixture}
\alias{DSD_Mixture}
\title{Mixes Data Points from Several Streams into a Single Stream}
\usage{
DSD_Mixture(..., prob = NULL)
}
\arguments{
\item{...}{\link{DSD} objects.}

\item{prob}{a numeric vector with the probability for each stream that the next point will
be drawn from that stream.}
}
\value{
Returns a \code{DSD_Mixture} object.(subclass of \link{DSD_R}, \link{DSD}).
}
\description{
This generator mixes multiple streams given specified probabilities.
The streams have to contain the same number of dimensions.
}
\examples{
# create data stream with three clusters in 2D
stream1 <- DSD_Gaussians(d = 2, k = 3)
stream2 <- DSD_UniformNoise(d = 2,  range = rbind(c(-.5, 1.5), c(-.5, 1.5)))

combinedStream <- DSD_Mixture(stream1, stream2, prob = c(.9, .1))
combinedStream

get_points(combinedStream, n = 20)
plot(combinedStream, n = 200)
}
\seealso{
Other DSD:
\code{\link[=DSD]{DSD()}},
\code{\link[=DSD_BarsAndGaussians]{DSD_BarsAndGaussians()}},
\code{\link[=DSD_Benchmark]{DSD_Benchmark()}},
\code{\link[=DSD_Cubes]{DSD_Cubes()}},
\code{\link[=DSD_Gaussians]{DSD_Gaussians()}},
\code{\link[=DSD_MG]{DSD_MG()}},
\code{\link[=DSD_Memory]{DSD_Memory()}},
\code{\link[=DSD_NULL]{DSD_NULL()}},
\code{\link[=DSD_ReadDB]{DSD_ReadDB()}},
\code{\link[=DSD_ReadStream]{DSD_ReadStream()}},
\code{\link[=DSD_Target]{DSD_Target()}},
\code{\link[=DSD_UniformNoise]{DSD_UniformNoise()}},
\code{\link[=DSD_mlbenchData]{DSD_mlbenchData()}},
\code{\link[=DSD_mlbenchGenerator]{DSD_mlbenchGenerator()}},
\code{\link[=DSF]{DSF()}},
\code{\link[=animate_data]{animate_data()}},
\code{\link[=close_stream]{close_stream()}},
\code{\link[=get_points]{get_points()}},
\code{\link[=plot.DSD]{plot.DSD()}},
\code{\link[=reset_stream]{reset_stream()}}
}
\author{
Michael Hahsler
}
\concept{DSD}
