% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Scale.R
\name{DSD_ScaleStream}
\alias{DSD_ScaleStream}
\title{Deprecated DSD_ScaleStream}
\usage{
DSD_ScaleStream(...)
}
\arguments{
\item{...}{arguments are passed on to \code{\link[=DSF_Scale]{DSF_Scale()}}.}
}
\value{
Produces a warning that \code{DSD_ScaleStream} is deprecated
and returns an object of class
\link{DSF_Scale} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Deprecated! Use \link{DSF_Scale} instead.
}
