% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_FeaturesSelection.R
\name{DSF_FeatureSelection}
\alias{DSF_FeatureSelection}
\title{Select Features for a Data Stream}
\usage{
DSF_FeatureSelection(dsd = NULL, features)
}
\arguments{
\item{dsd}{An object of class \link{DSD}.}

\item{features}{A character vector of feature (column) names or a numeric
vector of feature indices. All other features are removed. Special
information columns starting with \code{.} are not features.}
}
\value{
An object of class \code{DSF_FeatureSelection} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Selects features from a data stream using a supplied list of features.
}
\examples{
stream <- DSD_Gaussians(k = 3, d = 3)
get_points(stream, 3)

stream_2features <- DSF_FeatureSelection(stream, features = c("X1", "X3"))
stream_2features

get_points(stream_2features, n = 3)
}
\seealso{
Other DSF:
\code{\link[=DSF]{DSF()}},
\code{\link[=DSF_Convolve]{DSF_Convolve()}},
\code{\link[=DSF_Downsample]{DSF_Downsample()}},
\code{\link[=DSF_ExponentialMA]{DSF_ExponentialMA()}},
\code{\link[=DSF_Func]{DSF_Func()}},
\code{\link[=DSF_Scale]{DSF_Scale()}},
\code{\link[=DSF_dplyr]{DSF_dplyr()}}
}
\author{
Michael Hahsler
}
\concept{DSF}
