% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Filter.R
\name{DSF_Filter}
\alias{DSF_Filter}
\alias{filter_MA}
\alias{filter_hamming}
\alias{filter_sinc}
\title{Apply a Filter to a Data Stream}
\usage{
DSF_Filter(dsd, dim = NULL, kernel = NULL, na.rm = TRUE)

filter_MA(width)

filter_hamming(width)

filter_sinc(fc, fs, width = NULL, bw = NULL)
}
\arguments{
\item{dsd}{A object of class \link{DSD}.}

\item{dim}{columns to which the filter should be applied. Default is all.}

\item{kernel}{filter kernel as a vector of weights.}

\item{na.rm}{logical; should NAs be ignored?}

\item{width}{filter width.}

\item{fc}{cutoff frequency.}

\item{fs}{sampling frequency.}

\item{bw}{transition bandwidth.}
}
\value{
An object of class \code{DSF_Filter} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Applies a filter (i.e., a convolution with a filter kernel) to a data stream.
}
\details{
A filter kernel is a vector with kernel weights. A few filter are provided.
}
\examples{
data(presidents)

## create a data stream with two copies of president approval ratings.
## we will filter the second one and leave the first unfiltered.
stream <- DSD_Memory(data.frame(approval_orig = presidents, approval_MA = presidents))
plot(stream, dim = 1, n = 120, method = "ts")

## apply a moving average filter to dimension 1
filteredStream <- DSF_Filter(stream, kernel = filter_MA(5), dim = 2)
filteredStream

## resetting the filtered stream also resets the original stream
reset_stream(filteredStream)
get_points(filteredStream, n = 10)

reset_stream(filteredStream)
plot(filteredStream, dim = 1, n = 120, method = "ts")

## the second dimension is filtered
reset_stream(filteredStream)
plot(filteredStream, dim = 2, n = 120, method = "ts")

## look at different filters
plot(filter_MA(20), type = "l")
plot(filter_hamming(20), type = "l")
plot(filter_sinc(10, 100, width = 20), type = "l")

}
\seealso{
\link[stats:filter]{stats::filter}

Other DSF: 
\code{\link{DSF_Downsample}()},
\code{\link{DSF_Scale}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
