% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Scale.R
\name{DSF_Scale}
\alias{DSF_Scale}
\title{Scale a Data Stream}
\usage{
DSF_Scale(dsd = NULL, dim = NULL, center = TRUE, scale = TRUE, n = 100L)
}
\arguments{
\item{dsd}{An object of class \link{DSD} to scale.}

\item{dim}{Integer indices or names of the dimensions to scale. The default
is all dimensions.}

\item{center, scale}{Logical values or numeric vectors of scaling factors,
each with one value per selected column (see \link{scale}).}

\item{n}{Number of points used to estimate centering and scaling values.}
}
\value{
An object of class \code{DSF_Scale} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Converts an unscaled data stream into a scaled data stream.
}
\details{
If \code{center} or \code{scale} is logical, \code{DSF_Scale()} estimates the corresponding
values from \code{n} points in the stream using \link{scale} in \pkg{base}. Estimating
these values advances the stream by \code{n} points.
}
\section{Deprecated}{

\code{DSD_ScaleStream} is deprecated. Use \code{DSF_Scale} instead.
}

\examples{
stream <- DSD_Gaussians(k = 3, d = 2)
get_points(stream, 3)

# scale with manually calculated scaling factors
points <- get_points(stream, n = 100, info = FALSE)
center <- colMeans(points)
scale <- apply(points, MARGIN = 2, sd)

scaledStream <- stream \%>\% DSF_Scale(dim = c(1L, 2L), center = center, scale = scale)
colMeans(get_points(scaledStream, n = 100, info = FALSE))
apply(get_points(scaledStream, n = 100, info = FALSE), MARGIN = 2, sd)

# let DSF_Scale calculate the scaling factors from the first n points of the stream
scaledStream <- stream \%>\% DSF_Scale(dim = c(1L, 2L), n = 100)
colMeans(get_points(scaledStream, n = 100, info = FALSE))
apply(get_points(scaledStream, n = 100, info = FALSE), MARGIN = 2, sd)

## scale only X2
scaledStream <- stream \%>\% DSF_Scale(dim = "X2", n = 100)
colMeans(get_points(scaledStream, n = 100, info = FALSE))
apply(get_points(scaledStream, n = 100, info = FALSE), MARGIN = 2, sd)
}
\seealso{
\link{scale} in \pkg{base}

Other DSF:
\code{\link[=DSF]{DSF()}},
\code{\link[=DSF_Convolve]{DSF_Convolve()}},
\code{\link[=DSF_Downsample]{DSF_Downsample()}},
\code{\link[=DSF_ExponentialMA]{DSF_ExponentialMA()}},
\code{\link[=DSF_FeatureSelection]{DSF_FeatureSelection()}},
\code{\link[=DSF_Func]{DSF_Func()}},
\code{\link[=DSF_dplyr]{DSF_dplyr()}}
}
\author{
Michael Hahsler
}
\concept{DSF}
