% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_ScaleStream.R
\name{DSF_ScaleStream}
\alias{DSF_ScaleStream}
\title{Scale a Data Stream}
\usage{
DSF_ScaleStream(dsd, center = TRUE, scale = TRUE, n = 1000, reset = FALSE)
}
\arguments{
\item{dsd}{A object of class \link{DSD} that will be scaled.}

\item{center, scale}{logical or a numeric vector of length equal to the
number of columns used for centering/scaling (see function \link{scale}).}

\item{n}{The number of points used by \code{scale_stream()} to creating the centering/scaling}

\item{reset}{Try to reset the stream to its beginning after taking \code{n}
points for scaling.}
}
\value{
An object of class \code{DSF_ScaleStream} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Make an unscaled data stream into a scaled data stream.
}
\details{
\code{scale_stream()} estimates the values for centering and scaling (see
\link{scale} in \pkg{base}) using \code{n} points from the stream.
}
\examples{
stream <- DSD_Gaussians(k=3, d=3)
plot(stream)

# scale stream using 100 points
stream_scaled <- DSF_ScaleStream(stream, n=100)
stream_scaled

plot(stream_scaled)
}
\seealso{
\link{scale} in \pkg{base}

Other DSF: 
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
