% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFP.R
\name{DSFP}
\alias{DSFP}
\title{Abstract Class for Frequent Pattern Mining Algorithms for Data Streams}
\usage{
DSFP(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for frequent pattern mining algorithms for data streams.
Currently, \pkg{stream} does not implement frequent pattern mining
algorithms.
}
\examples{
DSFP()
}
\seealso{
\link{DST}
}
\author{
Michael Hahsler
}
