% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSO.R
\name{DSO}
\alias{DSO}
\alias{update.DSO}
\alias{get_points.DSO}
\alias{get_weights.DSO}
\title{Data Stream Operator Base Classes}
\usage{
DSO(...)

\method{update}{DSO}(object, dsd, n = 1, ...)

\method{get_points}{DSO}(x, ...)

\method{get_weights}{DSO}(x, ...)
}
\arguments{
\item{...}{Further arguments.}

\item{dsd}{a data stream object.}

\item{n}{the number of data points used for the update.}

\item{x, object}{a concrete implementation of DSO}
}
\description{
Abstract base classes for all DSO (Data Stream Operator) classes.
}
\details{
The \code{DSO} class cannot be instantiated, but it serve as a base
class from which other DSO classes inherit.

Data stream operators use \link{update} to process new data from the
\link{DSD} stream. The result of the operator can be obtained via \link{get_points}
and \link{get_weights} (if available).
}
\examples{
DSO()
}
\seealso{
Other DSO: 
\code{\link{DSO_Sample}()},
\code{\link{DSO_Window}()}
}
\author{
Michael Hahsler
}
\concept{DSO}
