% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSRegressor.R
\name{DSRegressor}
\alias{DSRegressor}
\title{Abstract Class for Data Stream Regressors}
\usage{
DSRegressor(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for data stream regressors. More implementations can be found in
package \pkg{streamMOA}.
}
\examples{
DSClassifier()
}
\seealso{
Other DST:
\code{\link[=DSAggregate]{DSAggregate()}},
\code{\link[=DSC]{DSC()}},
\code{\link[=DSClassifier]{DSClassifier()}},
\code{\link[=DSOutlier]{DSOutlier()}},
\code{\link[=DST]{DST()}},
\code{\link[=DST_SlidingWindow]{DST_SlidingWindow()}},
\code{\link[=DST_WriteStream]{DST_WriteStream()}},
\code{\link{evaluate}},
\code{\link{predict}},
\code{\link{stream_pipeline}},
\code{\link{update}}

Other DSRegressor:
\code{\link[=DSRegressor_SlidingWindow]{DSRegressor_SlidingWindow()}}
}
\author{
Michael Hahsler
}
\concept{DSRegressor}
\concept{DST}
