% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_EA.R
\docType{class}
\name{EA-class}
\alias{EA-class}
\alias{EA_R}
\title{Reference Class EA_R}
\description{
Reference class used for Reclustering using an evolutionary algorithm
}
\section{Fields}{

\describe{
\item{\code{crossoverRate}}{cross-over rate for the evolutionary algorithm}

\item{\code{mutationRate}}{mutation rate for the evolutionary algorithm}

\item{\code{populationSize}}{number of solutions that the evolutionary algorithm maintains}

\item{\code{k}}{number of macro-clusters}

\item{\code{generations}}{number of EA generations performed during reclustering}

\item{\code{data}}{micro-clusters to recluster}

\item{\code{weights}}{weights of the micro-clusters}

\item{\code{C}}{exposed C class}
}}


\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
