% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stream.R
\name{enrich_genes}
\alias{enrich_genes}
\title{Perform enrichment analysis for eRegulon genes against gene ontology (GO) terms
and KEGG pathways}
\usage{
enrich_genes(regs = NULL, dbs = c("GO", "KEGG"), org = "human")
}
\arguments{
\item{regs}{The list of enhancer regulons (eRegulons) or cell-type-specific eRegulons.}

\item{dbs}{The list of databases to run enrichment analysis, c("GO", "KEGG") by default.}

\item{org}{The organism, "human" by default.}
}
\description{
This function aims to calculate the enrichment of genes in each eRegulon against
gene ontology (GO) terms or KEGG pathways. Finally, this function returns a list of \code{data.frame}
objects, each of which represents one functional genomics databases, e.g., Biological Process,
Molecular Function, Cellular Component, and KEGG pathways (human or mouse).
}
\references{
Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022.
Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data.
bioRxiv, pp.2022-12.

Kuleshov, M. V., Jones, M. R., Rouillard, A. D., Fernandez, N. F., Duan, Q.,
Wang, Z., ... & Ma'ayan, A. (2016). Enrichr: a comprehensive gene set enrichment analysis web
server 2016 update. Nucleic acids research, 44(W1), W90-W97.
}
