% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_evoStream.R
\docType{class}
\name{evoStream_R-class}
\alias{evoStream_R-class}
\alias{evoStream_R}
\title{Reference Class evoStream_R}
\description{
Reference class mostly used to expose the C class object
}
\section{Fields}{

\describe{
\item{\code{C}}{exposed C class}
}}


\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
