% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reguome_tools.R
\name{find_TFBS}
\alias{find_TFBS}
\title{Annotate enhancers with TFs}
\usage{
find_TFBS(peaks, TFBS.list, org = "hg38", candidate.TFs = NULL)
}
\description{
Annotate enhancers with TFs
}
\keyword{internal}
