\name{get_copy}
\alias{get_copy}
\title{Create a Deep Copy of a DSC Object}

\description{
DSC objects contain reference classes or Java data structures 
(for MOA). Therefore, we provide a mechanism to create deep copies.
}

\usage{
	get_copy(x)
}
\arguments{
	\item{x}{The DSC object being copied.}
}

%\details{
%}

\value{
A deep copy of the original DSC.
}

\seealso{
	\code{\link{DSC}}
}

\author{
    Michael Hahsler
}


%\examples{
%}
