% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{get_coverage_plot}
\alias{get_coverage_plot}
\title{Generate coverage plot within an interval across different cell types}
\usage{
get_coverage_plot(
  object = NULL,
  region = NULL,
  features = NULL,
  links = FALSE,
  ranges.group.by = "seurat_cluster",
  peaks = FALSE,
  bigwig = NULL,
  region.highlight = NULL,
  idents = NULL,
  extend.upstream = 0,
  extend.downstream = 0,
  colors = NULL,
  size = 10
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{region}{A set of genomic coordinates to show. Can be a \code{GRanges} object,
a string encoding a genomic position, a gene name, or a vector of strings describing
the genomic coordinates or gene names to plot. If a gene name is supplied, annotations
must be present in the assay.}

\item{features}{A vector of features present in another assay to plot alongside
accessibility tracks (for example, gene names).}

\item{links}{Display links}

\item{ranges.group.by}{Grouping variable to color \code{ranges} by. Must be a variable present
in the metadata stored in the \code{ranges} genomic ranges. If NULL, do not color by any variable.}

\item{peaks}{Display peaks}

\item{bigwig}{List of bigWig file paths to plot data from. Files can be remotely hosted.
The name of each element in the list will determine the y-axis label given to the track.}

\item{region.highlight}{Region to highlight on the plot. Should be a \code{GRanges} object
containing the coordinates to highlight. By default, regions will be highlighted in grey.
To change the color of the highlighting, include a metadata column in the
GRanges object named "color" containing the color to use for each region.}

\item{idents}{Which identities to include in the plot. Default is all identities.}

\item{extend.upstream}{Number of bases to extend the region upstream.}

\item{extend.downstream}{Number of bases to extend the region downstream.}

\item{colors}{A list of colors to fill different genomic ranges}

\item{size}{The font size}
}
\description{
Generate coverage plot within an interval across different cell types
}
