% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{get_multi_hist}
\alias{get_multi_hist}
\title{Generate multiple histograms for two values}
\usage{
get_multi_hist(
  df = NULL,
  ylab = "Frequency",
  binwidth = NULL,
  size = 15,
  fill = "#63707E",
  hjust = 0,
  path = NULL,
  font.scale = 1,
  strip.face = "plain"
)
}
\description{
Generate multiple histograms for two values
}
\keyword{internal}
