% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{GiottoToSpark}
\alias{GiottoToSpark}
\title{Transform Giotto object to Spark}
\usage{
GiottoToSpark(object, ...)
}
\arguments{
\item{object}{An Giotto object}

\item{...}{Other arguments sell also \code{\link[SPARK]{CreateSPARKObject}}}
}
\value{
An SPARK object
}
\description{
Transform Giotto object to Spark
}
