% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Integration}
\alias{Integration}
\title{Integration single cell RNA-seq data and spatial single cell RNA-seq data}
\usage{
Integration(object, ...)
}
\arguments{
\item{object}{An giotto object}

\item{...}{Argument passed to other methods}
}
\description{
Integration single cell RNA-seq data and spatial single cell RNA-seq data
}
