% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stream.R
\name{intersect_enhancer_gene_relations_in_batch}
\alias{intersect_enhancer_gene_relations_in_batch}
\title{Calculate the precision, recall, and f-scores of the overlaps between
two lists of \code{GRanges} objects indicating enhancer-gene relations}
\usage{
intersect_enhancer_gene_relations_in_batch(
  link.pairs,
  ep.ll,
  only.overlap = FALSE,
  max.score = "precision"
)
}
\arguments{
\item{link.pairs}{The first list of \code{GRanges} objects saving enhancer-gene relations.}

\item{ep.ll}{The second list of \code{GRanges} objects saving enhancer-gene relations.}

\item{only.overlap}{Only consider the \code{GRanges} objects of which genes were overlapped against databases,
TRUE by default.}

\item{max.score}{Which score will be used to select the best query-hit pairs of \code{GRanges} objects,
"precision" by default.}
}
\value{
Returns a \code{data.frame} to indicate the query-hit pairs as well as precision, recall, and f-score.
The \code{data.frame} contains the following columns:
\item{EP}{The ID of overlapped enhancer-gene pairs in databases.}
\item{precision}{The precision of the overlaps between enhancer-gene relations between
enhancer regulons (eRegulons) and that in databases.}
\item{recall}{The recall of the overlaps between enhancer-gene relations between
eRegulons and that in databases.}
\item{fscore}{The f-score of the overlaps between enhancer-gene relations between
eRegulons and that in databases.}
}
\description{
This function has the same functionality as the function "intersect_enhancer_gene_relations".
The only difference is that this function aims to compare two lists of \code{GRanges} objects.
}
\references{
Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022.
Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data.
bioRxiv, pp.2022-12.

Gao, T., & Qian, J. (2020). EnhancerAtlas 2.0: an updated resource with enhancer
annotation in 586 tissue/cell types across nine species. Nucleic acids research, 48(D1), D58-D64.

Gao, T., Zheng, Z., Pan, Y., Zhu, C., Wei, F., Yuan, J., ... & Qian, J. (2022).
scEnhancer: a single-cell enhancer resource with annotation across hundreds of tissue/cell
types in three species. Nucleic acids research, 50(D1), D371-D379.
}
