% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stream.R
\name{intersect_peaks}
\alias{intersect_peaks}
\title{Calculate the p-values of overlaps between two \code{GRanges} objects}
\usage{
intersect_peaks(x, y, n.times = 100, alternative = "greater")
}
\arguments{
\item{x}{The first \code{GRanges} object or \code{data.frame}.}

\item{y}{The first \code{GRanges} object or \code{data.frame}.}

\item{n.times}{The number of times of permutation, 1000 by default.}

\item{alternative}{The direction of alternative hypothesis, "greater" by default.}
}
\value{
Return a numeric p-value.
}
\description{
Given two \code{GRanges} objects, this function calculates the overlaps between them.
Based on the calculated overlaps,
this function relies upon \code{regioneR} to perform permutation test to assess the significance of
overlaps between the two \code{GRanges} objects. Finally, a p-value will be calculated. Usually, we
perform comparison for the enhancer set of an eRegulon against a series of ChIP-seq peaks in the same tissues
or cell lines.
}
\references{
Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022.
Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data.
bioRxiv, pp.2022-12.

Gel, B., Díez-Villanueva, A., Serra, E., Buschbeck, M., Peinado, M. A., & Malinverni, R. (2016).
regioneR: an R/Bioconductor package for the association analysis of genomic regions based on permutation tests.
Bioinformatics, 32(2), 289-291.
}
