% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stream.R
\name{intersect_peaks_in_batch}
\alias{intersect_peaks_in_batch}
\title{Calculate the p-values of overlaps between two lists of \code{GRanges} objects}
\usage{
intersect_peaks_in_batch(x.ll, y.ll, n.times = 100)
}
\arguments{
\item{x.ll}{The first list of \code{GRanges} objects.}

\item{y.ll}{The first list of \code{GRanges} objects.}

\item{n.times}{The number of times of permutation, 1000 by default.}
}
\value{
Returns a \code{data.frame} composed of the IDs of significantly overlapped
\code{GRanges} objects and p-values.
}
\description{
This function has the same functionality as the function "intersect_peaks". The
only difference is that this function compares two lists of \code{GRanges} objects and returns
a \code{data.frame} composed of the IDs in each \code{GRanges} object as well as the p-values.
}
\references{
Li, Y., Ma, A., Wang, Y., Wang, C., Chen, S., Fu, H., Liu, B. and Ma, Q., 2022.
Enhancer-driven gene regulatory networks inference from single-cell RNA-seq and ATAC-seq data.
bioRxiv, pp.2022-12.

Gel, B., Díez-Villanueva, A., Serra, E., Buschbeck, M., Peinado, M. A., & Malinverni, R. (2016).
regioneR: an R/Bioconductor package for the association analysis of genomic regions based on permutation tests.
Bioinformatics, 32(2), 289-291.
}
