% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{link_peaks}
\alias{link_peaks}
\title{Link peaks to genes using \code{Signac} functions}
\usage{
link_peaks(
  object,
  peak.assay = "ATAC",
  expression.assay = "RNA",
  peak.slot = "counts",
  expression.slot = "data",
  method = "pearson",
  gene.coords = NULL,
  distance = 5e+05,
  min.distance = NULL,
  min.cells = 10,
  genes.use = NULL,
  n_sample = 200,
  pvalue_cutoff = 0.05,
  score_cutoff = 0.05,
  gene.id = FALSE,
  verbose = TRUE
)
}
\description{
Link peaks to genes using \code{Signac} functions
}
\keyword{internal}
