% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{LinksToGRanges}
\alias{LinksToGRanges}
\title{Link matrix to granges}
\usage{
LinksToGRanges(linkmat, gene.coords, sep = c("-", "-"))
}
\arguments{
\item{linkmat}{A sparse matrix with genes in the rows and peaks in the
columns}

\item{gene.coords}{Genomic coordinates for each gene}
}
\value{
Returns a GRanges object
}
\description{
Link matrix to granges
}
\keyword{internal}
