% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{overlap_peak_lst}
\alias{overlap_peak_lst}
\title{Overlap two lists of \code{GRanges} objects}
\usage{
overlap_peak_lst(lst1, lst2)
}
\description{
Overlap two lists of \code{GRanges} objects
}
\keyword{internal}
