% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.DSC}
\alias{plot.DSD}
\title{Plotting Data Stream Data and Clusterings}
\usage{
\method{plot}{DSC}(
  x,
  dsd = NULL,
  n = 500,
  col_points = NULL,
  col_clusters = c("red", "blue", "green"),
  weights = TRUE,
  scale = c(1, 5),
  cex = 1,
  pch = NULL,
  method = "pairs",
  dim = NULL,
  type = c("auto", "micro", "macro", "both", "all", "outliers"),
  assignment = FALSE,
  ...
)

\method{plot}{DSD}(
  x,
  n = 500,
  col = NULL,
  pch = NULL,
  ...,
  method = "pairs",
  dim = NULL,
  alpha = 0.6
)
}
\arguments{
\item{x}{the \link{DSD} or \link{DSC} object to be plotted.}

\item{dsd}{a \link{DSD} object to plot the data in the background if \code{x} is a \link{DSC}.}

\item{n}{number of plots taken from the dsd to plot.}

\item{weights}{the size of the symbols for micro- and macro-clusters
represents its weight.}

\item{scale}{range for the symbol sizes used.}

\item{cex}{size factor for symbols.}

\item{pch}{symbol type.}

\item{method}{method used for plotting: \code{"pairs"} (pairs plot), \code{"scatter"}
(scatter plot), \code{"pca"} (plot first 2 principal components), or \code{"ts"} (time series).}

\item{dim}{an integer vector with the dimensions to plot. If \code{NULL}
then for methods \code{pairs} and \code{"pca"} all dimensions are used and
for \code{"scatter"} the first two dimensions are plotted.}

\item{type}{Plot micro clusters (\code{type = "micro"}), macro clusters
(\code{type = "macro"}), both micro and macro clusters (\code{type = "both"}),
outliers(\code{type = "outliers"}), or everything together
(\code{type = "all"}). \code{type = "auto"} leaves to the class of DSC to
decide.}

\item{assignment}{logical; show assignment area of micro-clusters.}

\item{\dots}{further arguments are passed on to \link{plot} or \link{pairs} in
\pkg{graphics}.}

\item{col, col_points, col_clusters}{colors used for plotting.}

\item{alpha}{alpha shading used to plot the points.}
}
\description{
Methods to plot data stream data and clusterings.
}
\examples{
stream <- DSD_Gaussians(k=3, d=3)

## plot data
plot(stream, n=500)
plot(stream, method="pca", n=500)
plot(stream, method="scatter", dim=c(1,3), n=500)

## create and plot micro-clusters
dstream <- DSC_DStream(gridsize=0.1)
update(dstream, stream, 500)
plot(dstream)

## plot with data, projected on the first two principal components
## and dimensions 2 and 3
plot(dstream, stream)
plot(dstream, stream, method="pc")
plot(dstream, stream, dim=c(2,3))

## plot micro and macro-clusters
plot(dstream, stream, type="both")
}
\seealso{
Other DSC: 
\code{\link{DSC_BICO}()},
\code{\link{DSC_BIRCH}()},
\code{\link{DSC_DBSCAN}()},
\code{\link{DSC_DBSTREAM}()},
\code{\link{DSC_DStream}()},
\code{\link{DSC_EA}()},
\code{\link{DSC_Hierarchical}()},
\code{\link{DSC_Kmeans}()},
\code{\link{DSC_Macro}()},
\code{\link{DSC_Micro}()},
\code{\link{DSC_Reachability}()},
\code{\link{DSC_R}()},
\code{\link{DSC_Sample}()},
\code{\link{DSC_Static}()},
\code{\link{DSC_TwoStage}()},
\code{\link{DSC_Window}()},
\code{\link{DSC_evoStream}()},
\code{\link{DSC}()},
\code{\link{animate}},
\code{\link{evaluate}()},
\code{\link{get_assignment}()},
\code{\link{microToMacro}()},
\code{\link{prune_clusters}()},
\code{\link{read_saveDSC}},
\code{\link{recluster}()}

Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_ReadCSV}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{DSF}()},
\code{\link{animate}},
\code{\link{get_points}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
Michael Hahsler
}
\concept{DSC}
\concept{DSD}
