% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_dotplot_heatmap}
\alias{plot_dotplot_heatmap}
\title{Get the dotplot-heatmap for cell-type-specific enhancer regulons (eRegulons)}
\usage{
plot_dotplot_heatmap(
  obj = NULL,
  cts.en.regs = NULL,
  n.regs = 5,
  celltype.col = "seurat_clusters",
  dot = "AC",
  peak.assay = "ATAC",
  ht.colours = c((grDevices::colorRampPalette(c("#99D5FF", "#E8F9FD")))(3), "#fffaf7",
    (grDevices::colorRampPalette(c("#ffe4e1", "#ee7261")))(3)),
  ht.breaks = NULL,
  font.size = 15,
  text.font = "Arial",
  angle = 30,
  celltypes = NULL,
  vjust = 1,
  hjust = 1,
  celltype.colours = NULL
)
}
\arguments{
\item{obj}{A \code{Seurat} object, NULL by default}

\item{cts.en.regs}{The list of cell-type-specific enhancer regulons (eRegulons)}

\item{n.regs}{The number of top-ranked eRegulons for plotting, 5 by default}

\item{celltype.col}{The metadata column indicating cell types, "seurat_clusters" by default}

\item{dot}{What dot sizes represent, "AC" by default}

\item{peak.assay}{The assay denoting chromatin accessibility, "ATAC" by default}

\item{ht.colours}{Color list used to define the gradients of the heatmap,
c(
grDevices::colorRampPalette(c("#99D5FF", "#E8F9FD"))(3),
"#fffaf7",
grDevices::colorRampPalette(c("#ffe4e1", "#ee7261"))(3)
)
by default}

\item{ht.breaks}{The breaks marked on the legend of heatmap colors}

\item{font.size}{The font size of legends}

\item{text.font}{The text font of legends}

\item{angle}{The angle of labels of the x-axis in the heatmap, 30 by default}

\item{celltypes}{The list of cell types arranged in an order defined by users}

\item{vjust}{Adjustment of positions in the vertical direction, 1 by default}

\item{hjust}{Adjustment of positions in the horizontal direction, 1 by default}

\item{celltype.colours}{The list of colors of cell types in the row annotations}
}
\description{
Get the dotplot-heatmap for cell-type-specific enhancer regulons (eRegulons)
}
