% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_eGRN}
\alias{plot_eGRN}
\title{Plot an enhancer-driven gene regulatory network (eGRN)}
\usage{
plot_eGRN(
  en.grn = NULL,
  obj = NULL,
  peak.assay = "ATAC",
  n.links = 20,
  title = "temp",
  path = "./",
  format = ".png",
  n.TFs = 15,
  layout = "fr",
  max.links = 5000,
  ifSaveRda = FALSE,
  TFs = NULL
)
}
\arguments{
\item{en.grn}{The enhancer gene regulatory network (eGRN) composed of \code{GRanges}, cell type, and cells}

\item{obj}{A \code{Seurat} object}

\item{peak.assay}{The scATAC-seq assay}

\item{n.links}{The cutoff of the number of TF-enhancer-gene relations for each TF to showcase.
We retain the top-ranked relations in decreasing order of the number of cells, where the genes including the
ones encoding the TFs are expressed, while the enhancers are accessible, 50 by default}

\item{title}{The title of the eGRN to be plotted, "temp" by default}

\item{path}{The path to save the image file, "./" by default}

\item{format}{The format of the image file, ".png" by default}

\item{n.TFs}{The maximum number of top-ranked TFs to generate eGRN plot, 10 by default}

\item{layout}{The layout for plotting networks, "fr" by default}

\item{ifSaveRda}{Whether save the \code{igraph} objects in a qsdave file}

\item{TFs}{The list of TFs for plotting}
}
\description{
Plot an enhancer-driven gene regulatory network (eGRN)
}
\keyword{internal}
