% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{plot_stats_genes_and_enhs}
\alias{plot_stats_genes_and_enhs}
\title{Generate multiple histograms for :
\enumerate{
\item The number of enhancers linked to each gene
\item The rank of enhancers linked to each gene
}}
\usage{
plot_stats_genes_and_enhs(
  obj = NULL,
  peak.assay = "ATAC",
  en.regs = NULL,
  binwidth = NULL,
  size = 15,
  fill = "#63707E",
  hjust = 0,
  font.scale = 1,
  strip.face = "plain"
)
}
\arguments{
\item{obj}{The \code{Seurat} object to identify enhancer regulons (eRegulons)}

\item{en.regs}{The list of eRegulons}
}
\value{
Return a \code{ggplot} object
}
\description{
Generate multiple histograms for :
\enumerate{
\item The number of enhancers linked to each gene
\item The rank of enhancers linked to each gene
}
}
