% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{prepare_coverage_plot}
\alias{prepare_coverage_plot}
\title{Prepare TF-enhancer-gene relations to generate coverage plots}
\usage{
prepare_coverage_plot(
  object = NULL,
  links = NULL,
  highlight.colors = NULL,
  peak.assay = "ATAC",
  highlight.width = 500
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{links}{A list of \code{GRanges} object to save the TF-enhancer-gene relations}

\item{highlight.colors}{The colors of regions to highlight}

\item{peak.assay}{The scATAC-seq assay}

\item{highlight.width}{The width of the regions to highlight, i.e., TF binding sites}
}
\description{
Prepare TF-enhancer-gene relations to generate coverage plots
}
