% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{SFP_seeding}
\alias{SFP_seeding}
\title{Seeding based on Steiner forest problem (SFP) model}
\usage{
SFP_seeding(
  motif.obj = NULL,
  G.list,
  obj.list,
  bound.TFs,
  binding.CREs,
  block.list,
  TFGene.pairs,
  rna.dis,
  atac.dis,
  KL = "min.exp",
  P = NULL,
  Q = NULL,
  score.cutoff = 1,
  TOP_TFS = Inf,
  ifWeighted = TRUE,
  quantile.cutoff = 4,
  peak.assay = "ATAC",
  ifPutativeTFs = FALSE,
  padj.cutoff = 0.05
)
}
\description{
Seeding based on Steiner forest problem (SFP) model
}
\keyword{internal}
