% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatcluster.R, R/generics.R
\name{KmeansCluster}
\alias{KmeansCluster}
\alias{HmrfCluster}
\alias{SpatCluster}
\title{Kmeans methods to calculate spatial gene expression pattern}
\usage{
KmeansCluster(
  object,
  cellType = "leiden",
  method = "maximum",
  k = 10,
  centers = 6,
  iter.max = 300,
  nstart = 10,
  ...
)

HmrfCluster(
  object,
  knn = 5,
  expressValue = "scaled",
  spatialGenes = NULL,
  outputFolder = NULL,
  kdomain = 5,
  ...
)

SpatCluster(object, ...)
}
\arguments{
\item{object}{An Giotto object}

\item{cellType}{Column in meta to assign cell-type-annotation for spatial clustering. Default as "leiden" when external annotation is missing.}

\item{method}{The distance measure to be used. This must be one 
of "euclidean", "maximum", "manhattan", "canberra", "binary" or 
"minkowski". Any unambiguous substring can be given.}

\item{k}{Defines k for the k-nearest neighbor algorithm}

\item{centers}{The number of clusters}

\item{iter.max}{The maximum number of iterations allowed}

\item{nstart}{If centers is a number, how many random sets should be chosen}

\item{...}{Argument passed to other methods}

\item{knn}{number of nearest neighbors based on physical distance}

\item{expressValue}{Expression values to use (default [scaled])}

\item{spatialGenes}{Spatial genes to use for HMRF}

\item{outputFolder}{Output folder to save results}

\item{kdomain}{Number of HMRF domains, default 5}
}
\value{
giotto object
}
\description{
Kmeans methods to calculate spatial gene expression pattern

HmrfCluster

Cluster analysis with spatial single RNA-seq data
}
