% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\name{term2gene}
\alias{term2gene}
\title{TERM2GENE}
\usage{
term2gene(Geneformat = Geneformat, Species = Species, category = category)
}
\arguments{
\item{Geneformat}{Format of gene name in count matrix, one of "ensembl_gene", "gene_symbol".}

\item{Species}{Species name, such as Homo sapiens or Mus musculus.}

\item{category}{MSigDB collection abbreviation, such as H or C1.}
}
\description{
TERM2GENE
}
