% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize.r
\name{spotPOPplot}
\alias{spotPOPplot}
\alias{Scoreplot}
\alias{QCplot}
\alias{scHOTplot}
\alias{SEplot}
\alias{dot_plot}
\alias{cellphondedb_dotplot}
\alias{cellphoneDBplot}
\title{spotPOPplot}
\usage{
spotPOPplot(
  object,
  outputFolder = NULL,
  title = "CellPOP",
  meta = object@cell_metadata$leiden,
  legend = "leiden",
  savePlot = FALSE,
  ...
)

Scoreplot(
  x,
  y,
  outputFolder = NULL,
  title = "Spatplot",
  value,
  legend = "score",
  savePlot = FALSE,
  ...
)

QCplot(object, outputFolder = NULL, ...)

scHOTplot(
  scHOTobj,
  outputFolder = NULL,
  title = "gene-gene coexpression",
  savePlot = FALSE,
  genes = NULL,
  ...
)

SEplot(
  diffgenes = diffgenes,
  topgenes = 100,
  method = "Integrated",
  outputFolder = NULL,
  savePlot = TRUE,
  Species = NULL,
  category = "C2",
  Geneformat = "gene_symbol",
  ...
)

dot_plot(
  interaction,
  selected_rows = NULL,
  selected_columns = NULL,
  means_separator = "\\t",
  pvalues_separator = "\\t",
  ...
)

cellphondedb_dotplot(Lcelltype = 1, Rcelltype = 2, interaction, top = 10)

cellphoneDBplot(
  interaction = interaction,
  outputFolder = NULL,
  save.plot = FALSE,
  ...
)
}
\arguments{
\item{object}{A giotto object}

\item{outputFolder}{Output folder to save results.}

\item{title}{A character for the title of plot.}

\item{meta}{A discrete vector to assign color for each spot, e.g., cell type annotation, HMRF result, expression of a certain gene.}

\item{legend}{A character of feature to divide up cell into sub-populations, e.g. "K-means", "HMRF", "cell type".}

\item{savePlot}{Boolean, whether to save plot.}

\item{...}{other arguments}

\item{x}{Numeric, x-coordinate of spatial data.}

\item{y}{Numeric, y-coordinate of spatial data.}

\item{value}{A vector to assign color for each spot, e.g., cell type annotation, HMRF result, expression of a certain gene.}

\item{scHOTobj}{A scHOT object}

\item{genes}{A gene pair in scHOT object with weighted higher order statistc, linked by "_" or taken as a vector.}

\item{diffgenes}{A list out of SE-genes calculation.}

\item{topgenes}{Integer; set how many top SE-genes to take, default as 100.}

\item{method}{A character to select the SE method for downstream analysis, one of "Binspect","silhouetteRank","SPARK","Integrated".}

\item{Species}{Species name, more details in \code{\link[msigdbr]{msigdbr_species}}}

\item{category}{MSigDB collection abbreviation, such as H or C1.}

\item{Geneformat}{Format of gene name in count matrix, one of "ensembl_gene", "gene_symbol".}

\item{interaction}{A list out of cellphoneDB calculation.}

\item{selected_rows}{selected clusters of cellphoneDB results}

\item{selected_columns}{selected genepairs of cellphoneDB results}

\item{means_separator}{separator of means value file}

\item{pvalues_separator}{separator of p-value file}

\item{Lcelltype}{}

\item{Rcelltype}{}

\item{top}{Number of top pairs to plot.}

\item{save.plot}{directly save the plot [boolean].}
}
\value{
ggplot object

ggplot object

ggplot object

ggplot object
}
\description{
Visualize the spatial data with a continuous value in each spot, e.g. expression, scHOT score.
}
