% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD.R
\name{close_stream}
\alias{close_stream}
\title{Close a Data Stream}
\usage{
close_stream(dsd)
}
\arguments{
\item{dsd}{An object of class a subclass of \link{DSD} which implements a
reset function.}
}
\description{
Close a data stream that needs closing (e.g., a file or a connection).
}
\details{
\code{close_stream()} is implemented for:
\itemize{
\item \link{DSD}
\item \link{DSD_ReadCSV}
\item \link{DSD_ReadDB}
\item \link{DSD_ReadStream}
\item \link{DSF}
}
}
\seealso{
Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Combination}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_NULL}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_ReadStream}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{DSF}()},
\code{\link{animate_data}()},
\code{\link{get_points}()},
\code{\link{plot.DSD}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
Michael Hahsler
}
\concept{DSD}
