% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_EA.R
\name{DSC_EA}
\alias{DSC_EA}
\title{Evolutionary Algorithm}
\usage{
DSC_EA(k, generations = 2000, crossoverRate = 0.8, mutationRate = 0.001,
  populationSize = 100)
}
\arguments{
\item{k}{number of macro-clusters}

\item{generations}{number of EA generations performed during reclustering}

\item{crossoverRate}{cross-over rate for the evolutionary algorithm}

\item{mutationRate}{mutation rate for the evolutionary algorithm}

\item{populationSize}{number of solutions that the evolutionary algorithm maintains}
}
\description{
Reclustering using an evolutionary algorithm.
This approach is used by \code{evoStream} but can be used for all micro-clusters.
The evolutionary algorithm uses existing clustering solutions and creates small variations of them by combining and randomly modfiying them.
The modified solutions can yield better partitions and thus can improve the clustering over time.
The evolutionary algorithm is incremental, which allows to improve existing macro-clusters instead of recomputing them every time.
}
\examples{
stream <- DSD_Gaussians(k = 3, d = 2)
dbstream <- DSC_DBSTREAM(r=0.1)
EA <- DSC_EA(k=3, generations=2000)

two <- DSC_TwoStage(dbstream, EA)
update(two, stream, n=1200)
plot(two, stream, type="both")

update(dbstream, stream, n = 1200)
recluster(EA, dbstream)
plot(EA, stream)


}
\author{
Matthias Carnein \email{Matthias.Carnein@uni-muenster.de}
}
