% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_evoStream.R
\name{DSC_evoStream}
\alias{DSC_evoStream}
\title{evoStream - Evolutionary Stream Clustering}
\usage{
DSC_evoStream(r, lambda = 0.001, tgap = 100, k = 2,
  crossoverRate = 0.8, mutationRate = 0.001, populationSize = 100,
  initializeAfter = 2 * k, incrementalGenerations = 1,
  reclusterGenerations = 1000)
}
\arguments{
\item{r}{radius threshold for micro-cluster assignment}

\item{lambda}{decay rate}

\item{tgap}{time-interval between outlier detection and clean-up}

\item{k}{number of macro-clusters}

\item{crossoverRate}{cross-over rate for the evolutionary algorithm}

\item{mutationRate}{mutation rate for the evolutionary algorithm}

\item{populationSize}{number of solutions that the evolutionary algorithm maintains}

\item{initializeAfter}{number of micro-cluster required for the initialization of the evolutionary algorithm.}

\item{incrementalGenerations}{number of EA generations performed after each observation}

\item{reclusterGenerations}{number of EA generations performed during reclustering}
}
\description{
Stream clustering algorithm based on evolutionary optimization.
The online component uses a simplified version of \code{DBSTREAM} to generate micro-clusters.
The micro-clusters are then incrementally reclustered using an evloutionary algorithm.
Evolutionary algorithms create slight variations by combining and randomly modifying existing solutions.
By iteratively selecting better solutions, an evolutionary pressure is created which improves the clustering over time.
Since the evolutionary algorithm is incremental, it is possible to apply it between observations, e.g. in the idle time of the stream.
Whenever there is idle time, we can call the \code{recluster} function of the reference class to improve the macro-clusters (see example).
The evolutionary algorithm can also be applied as a traditional reclustering step, or a combination of both.
In addition, this implementation also allows to evaluate a fixed number of generations after each observation.
}
\examples{
stream <- DSD_Memory(DSD_Gaussians(k = 3, d = 2), 1000)

## init evoStream
evoStream <- DSC_evoStream(r=0.05, k=3, incrementalGenerations=1, reclusterGenerations=1000)

## insert observations
update(evoStream, stream, n = 1000)

## micro clusters
get_centers(evoStream, type="micro")

## micro weights
get_weights(evoStream, type="micro")

## macro clusters
get_centers(evoStream, type="macro")

## macro weights
get_weights(evoStream, type="macro")

## plot result
reset_stream(stream)
plot(evoStream, stream, type = "both")

## if we have time, evaluate additional generations. This can be called at any time, also between observations.
## by default, 1 generation is evaluated after each observation and 1000 generations during reclustering (parameters)
evoStream$RObj$recluster(2000)

## plot improved result
reset_stream(stream)
plot(evoStream, stream, type="both")

## get assignment of micro to macro clusters
microToMacro(evoStream)

}
\references{
Carnein M. and Trautmann H. (2018), "evoStream - Evolutionary Stream Clustering Utilizing Idle Times", Big Data Research.
}
\author{
Matthias Carnein \email{Matthias.Carnein@uni-muenster.de}
}
