% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_Hierarchical.R
\name{DSC_Hierarchical}
\alias{DSC_Hierarchical}
\title{Hierarchical Micro-Cluster Reclusterer}
\usage{
DSC_Hierarchical(
  k = NULL,
  h = NULL,
  method = "complete",
  min_weight = NULL,
  description = NULL
)
}
\arguments{
\item{k}{The number of desired clusters.}

\item{h}{Height where to cut the dendrogram.}

\item{method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of \code{"ward"}, \code{"single"}, \code{"complete"}, "\verb{average"},
\code{"mcquitty"}, \code{"median"} or \code{"centroid"}.}

\item{min_weight}{micro-clusters with a weight less than this will be
ignored for reclustering.}

\item{description}{optional character string to describe the clustering method.}
}
\value{
A list of class \link{DSC}, \link{DSC_R}, \link{DSC_Macro}, and
\code{DSC_Hierarchical}. The list contains the following items:

\item{description}{The name of the algorithm in the DSC object.}
\item{RObj}{The underlying R object.}
}
\description{
Macro Clusterer.
Implementation of hierarchical clustering to recluster a set of
micro-clusters.
}
\details{
Please refer to \code{\link[=hclust]{hclust()}} for more details on the behavior of the
algorithm.

\code{\link[=update]{update()}} and \code{\link[=recluster]{recluster()}} invisibly return the assignment of the data points
to clusters.

\strong{Note} that this clustering cannot be updated iteratively and every time it is
used for (re)clustering, the old clustering is deleted.
}
\examples{
# Cassini dataset
stream <- DSD_mlbenchGenerator("cassini")

# Use hierarchical clustering to recluster micro-clusters
dbstream <- DSC_DBSTREAM(r = .05)
update(dbstream, stream, 500)

# reclustering using single-link and specifying k
hc <- DSC_Hierarchical(k = 3, method = "single")
recluster(hc, dbstream)
hc
plot(hc, stream, type = "both")

# reclustering by specifying height
hc <- DSC_Hierarchical(h = .2, method = "single")
recluster(hc, dbstream)
hc
plot(hc, stream, type = "both")

# For comparison we use hierarchical clustering directly on the data
# Note: hierarchical clustering is not a data stream clustering algorithm!
hc <- DSC_Hierarchical(k = 3, method = "single")
assignment <- update(hc, stream, 500)
plot(hc, stream)

head(assignment)
}
\seealso{
Other DSC_Macro: 
\code{\link{DSC_DBSCAN}()},
\code{\link{DSC_EA}()},
\code{\link{DSC_Kmeans}()},
\code{\link{DSC_Macro}()},
\code{\link{DSC_Reachability}()}
}
\author{
Michael Hahsler
}
\concept{DSC_Macro}
