% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_Macro.R
\name{DSC_Macro}
\alias{DSC_Macro}
\title{Abstract Class for Macro Clusterers}
\usage{
DSC_Macro(...)
}
\arguments{
\item{...}{further arguments.}
}
\description{
Abstract class for all DSC Macro Clusterers. \code{DSC_Macro} cannot be instantiated.
}
\details{
\code{DSC_Macro} provide \link{microToMacro} that returns the assignment of Micro-cluster IDs to Macro-cluster IDs.
}
\seealso{
\link{DSC}
}
\author{
Michael Hahsler
}
