\name{DSC_Marco}
\alias{DSC_Macro}
\title{Abstract Class for Macro Clusterers}
\description{Abstract class for all DSC Macro Clusterers.}

\details{
\code{DSC_Macro} cannot be instantiated. Calling \code{DSC_Macro()}
results in an error.
}

\seealso{
	\code{\link{DSC}}
}

\author{
    Michael Hahsler
}

