% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_Micro.R
\name{DSC_Micro}
\alias{DSC_Micro}
\title{Abstract Class for Micro Clusterers}
\usage{
DSC_Micro(...)
}
\arguments{
\item{...}{further arguments.}
}
\description{
Abstract class for all DSC Micro Clusterers.
}
\details{
\code{DSC_Micro} cannot be instantiated.
}
\seealso{
\link{DSC}

Other DSC: 
\code{\link{DSC_BICO}()},
\code{\link{DSC_BIRCH}()},
\code{\link{DSC_DBSCAN}()},
\code{\link{DSC_DBSTREAM}()},
\code{\link{DSC_DStream}()},
\code{\link{DSC_EA}()},
\code{\link{DSC_Hierarchical}()},
\code{\link{DSC_Kmeans}()},
\code{\link{DSC_Macro}()},
\code{\link{DSC_Reachability}()},
\code{\link{DSC_R}()},
\code{\link{DSC_Sample}()},
\code{\link{DSC_Static}()},
\code{\link{DSC_TwoStage}()},
\code{\link{DSC_Window}()},
\code{\link{DSC_evoStream}()},
\code{\link{DSC}()},
\code{\link{animate}},
\code{\link{evaluate}()},
\code{\link{get_assignment}()},
\code{\link{microToMacro}()},
\code{\link{plot}()},
\code{\link{prune_clusters}()},
\code{\link{read_saveDSC}},
\code{\link{recluster}()}
}
\author{
Michael Hahsler
}
\concept{DSC}
