% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_Micro.R
\name{DSC_Micro}
\alias{DSC_Micro}
\title{Abstract Class for Micro Clusterers}
\usage{
DSC_Micro(...)
}
\arguments{
\item{...}{further arguments.}
}
\description{
Abstract class for all DSC Micro Clusterers.
}
\details{
\code{DSC_Micro} cannot be instantiated.
}
\seealso{
\link{DSC}
}
\author{
Michael Hahsler
}
