% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_TwoStage.R
\name{DSC_TwoStage}
\alias{DSC_TwoStage}
\title{TwoStage Clustering Process}
\usage{
DSC_TwoStage(micro, macro)
}
\arguments{
\item{micro}{Clustering algorithm used in the online stage
(\code{DSC_micro})}

\item{macro}{Clustering algorithm used for reclustering in the offline stage
(\code{DSC_macro})}
}
\value{
An object of class \code{DSC_TwoStage} (subclass of \code{DSC},
\code{DSC_Macro}).
}
\description{
Combines a micro and a macro clustering algorithm into a single process.
}
\details{
\code{update()} runs the micro-clustering stage and only when macro cluster
centers/weights are requested, then the offline stage reclustering is
automatically performed.
}
\examples{

stream <- DSD_Gaussians(k=3)

# Create a clustering process that uses a window for the online stage and
# k-means for the offline stage (reclustering)
win_km <- DSC_TwoStage(
  micro=DSC_Window(horizon=100),
  macro=DSC_Kmeans(k=3)
  )
win_km

update(win_km, stream, 200)
win_km
plot(win_km, stream, type="both")
evaluate(win_km, stream, assign="macro")

}
\seealso{
Other DSC: 
\code{\link{DSC_BICO}()},
\code{\link{DSC_BIRCH}()},
\code{\link{DSC_DBSCAN}()},
\code{\link{DSC_DBSTREAM}()},
\code{\link{DSC_DStream}()},
\code{\link{DSC_EA}()},
\code{\link{DSC_Hierarchical}()},
\code{\link{DSC_Kmeans}()},
\code{\link{DSC_Macro}()},
\code{\link{DSC_Micro}()},
\code{\link{DSC_Reachability}()},
\code{\link{DSC_R}()},
\code{\link{DSC_Sample}()},
\code{\link{DSC_Static}()},
\code{\link{DSC_Window}()},
\code{\link{DSC_evoStream}()},
\code{\link{DSC}()},
\code{\link{animate}},
\code{\link{evaluate}()},
\code{\link{get_assignment}()},
\code{\link{microToMacro}()},
\code{\link{plot}()},
\code{\link{prune_clusters}()},
\code{\link{read_saveDSC}},
\code{\link{recluster}()}
}
\author{
Michael Hahsler
}
\concept{DSC}
