% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSClassifier.R
\name{DSClassifier}
\alias{DSClassifier}
\title{Abstract Class for Data Stream Classifiers}
\usage{
DSClassifier(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for data stream classifiers. Currently, \pkg{stream} does not
implement classification algorithms. Implementations can be found in
package \pkg{streamMOA}.
}
\examples{
DSClassifier()
}
\seealso{
Other DST: 
\code{\link{DSAggregate}()},
\code{\link{DSC}()},
\code{\link{DSF_Scale}()},
\code{\link{DSOutlier}()},
\code{\link{DST_Runner}()},
\code{\link{DST_WriteStream}()},
\code{\link{DST}()},
\code{\link{evaluate}},
\code{\link{predict}()},
\code{\link{update}()}
}
\author{
Michael Hahsler
}
\concept{DST}
