% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSClassifier.R
\name{DSClassifier}
\alias{DSClassifier}
\title{Abstract Class for Data Stream Classifiers}
\usage{
DSClassifier(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for data stream classifiers. More implementations can be found in
package \pkg{streamMOA}.
}
\examples{
DSClassifier()
}
\seealso{
Other DST: 
\code{\link{DSAggregate}()},
\code{\link{DSC}()},
\code{\link{DSOutlier}()},
\code{\link{DSRegressor}()},
\code{\link{DST}()},
\code{\link{DST_SlidingWindow}()},
\code{\link{DST_WriteStream}()},
\code{\link{evaluate}},
\code{\link{predict}()},
\code{\link{stream_pipeline}},
\code{\link{update}()}

Other DSClassifier: 
\code{\link{DSClassifier_SlidingWindow}()}
}
\author{
Michael Hahsler
}
\concept{DSClassifier}
\concept{DST}
