% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSClassify.R
\name{DSClassify}
\alias{DSClassify}
\title{Abstract Class for Data Stream Classifiers}
\usage{
DSClassify(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract class for data stream classifiers. Currently, \pkg{stream} does not
implement classification algorithms.
}
\examples{
DSClassify()
}
\seealso{
\link{DST}

Other DST: 
\code{\link{DSAggregate}()},
\code{\link{DSC}()},
\code{\link{DSOutlier}()},
\code{\link{DST}()},
\code{\link{predict}()},
\code{\link{update}()}
}
\author{
Michael Hahsler
}
\concept{DST}
