% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_Benchmarks.R
\name{DSD_Benchmark}
\alias{DSD_Benchmark}
\title{Data Stream Generator for Benchmark Data}
\usage{
DSD_Benchmark(i = 1)
}
\arguments{
\item{i}{number of the benchmark.}
}
\value{
Returns a \code{DSD} object.
}
\description{
A data stream generator that generates several dynamic streams indented to
be benchmarks to compare data stream clustering algorithms.
}
\details{
Currently available benchmarks are 1 and 2.
}
\examples{

stream <- DSD_Benchmark(i=1)
\dontrun{
animate_data(stream, n=10000, horizon=100, xlim=c(0,1), ylim=c(0,1))
}

}
\seealso{
\code{\link{DSD}}

Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_ReadCSV}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{DSF}()},
\code{\link{animate}},
\code{\link{get_points}()},
\code{\link{plot}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
Michael Hahsler
}
\concept{DSD}
