% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_mlbenchGenerator.R
\name{DSD_mlbenchGenerator}
\alias{DSD_mlbenchGenerator}
\title{mlbench Data Stream Generator}
\usage{
DSD_mlbenchGenerator(method, ...)
}
\arguments{
\item{method}{The name of the mlbench data generator.}

\item{...}{Parameters for the mlbench data generator.}
}
\value{
Returns a \code{DSD_mlbenchGenerator} object (subclass of
\code{DSD_R}, \code{DSD}) which is a list of the defined parameters. The
parameters are either passed in from the function or created internally.
They include:

\item{description}{The name of the class of the DSD object.}
\item{method}{The name of the mlbench data generator.} \item{variables}{The
variables for the mlbench data generator.}
}
\description{
A data stream generator class that interfaces data generators found in
mlbench.
}
\details{
The \code{DSD_mlbenchGenerator} class is designed to be a wrapper class for
data created by data generators in the mlbench library.

Call \code{DSD_mlbenchGenerator} with missing method to get a list of
available methods.
}
\examples{

stream <- DSD_mlbenchGenerator(method="cassini")

plot(stream, n=500)

}
\seealso{
\code{\link{DSD}}

Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_ReadCSV}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_ScaleStream}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD}()},
\code{\link{animate}},
\code{\link{get_points}()},
\code{\link{plot}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
John Forrest
}
\concept{DSD}
