% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_NULL.R
\name{DSD_NULL}
\alias{DSD_NULL}
\title{Placeholder for a DSD Stream}
\usage{
DSD_NULL()
}
\value{
Returns a \code{DSD_NULL} object (subclass of \link{DSD}).
}
\description{
Placeholder for a \link{DSD}. \code{DSD_NULL} does not produce points and creates an error
for \code{\link[=get_points]{get_points()}}.
}
\examples{
nullstream <- DSD_NULL()
nullstream

## This will produce an error
\dontrun{
get_points(nullstream)}
}
\seealso{
Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_Mixture}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_ReadStream}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{DSF}()},
\code{\link{animate_data}()},
\code{\link{close_stream}()},
\code{\link{get_points}()},
\code{\link{plot.DSD}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
Michael Hahsler
}
\concept{DSD}
