% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF.R
\name{DSF}
\alias{DSF}
\title{Data Stream Filter Base Classes}
\usage{
DSF(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract base classes for all data stream filter (DSF) classes. Data stream filters transform a data stream (\link{DSD}).
}
\details{
The \code{DSF} class cannot be instantiated, but it serve as a base
class from which other DSF classes inherit.

Data stream filters transform a \link{DSD} data stream.
DSF implementations inherit from \link{DSD} and have the same basic interface.
}
\examples{
DSF()
}
\seealso{
Other DSF: 
\code{\link{DSF_ScaleStream}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
