% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Convolve.R
\name{DSF_Convolve}
\alias{DSF_Convolve}
\alias{filter_MA}
\alias{filter_hamming}
\alias{filter_diff}
\alias{filter_sinc}
\title{Apply a Filter to a Data Stream}
\usage{
DSF_Convolve(dsd, dim = NULL, kernel = NULL, na.rm = TRUE)

filter_MA(width)

filter_hamming(width)

filter_diff(lag)

filter_sinc(fc, fs, width = NULL, bw = NULL)
}
\arguments{
\item{dsd}{A object of class \link{DSD}.}

\item{dim}{columns to which the filter should be applied. Default is all.}

\item{kernel}{filter kernel as a vector of weights.}

\item{na.rm}{logical; should NAs be ignored?}

\item{width}{filter width.}

\item{fc}{cutoff frequency.}

\item{fs}{sampling frequency.}

\item{bw}{transition bandwidth.}
}
\value{
An object of class \code{DSF_Convolve} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Applies a filter (i.e., a convolution with a filter kernel) to a data stream.
}
\details{
A filter kernel is a vector with kernel weights. A few filter are provided.

Note about missing values: \code{na.rm} can be used to remove missing values (default behavior).
For \code{filter_diff()} removing missing values can be problematic and should not be used.
}
\examples{
data(presidents)

## create a data stream with two copies of president approval ratings.
## we will filter the second one and leave the first unfiltered.
stream <- data.frame(
    approval_orig = presidents,
    approval_MA = presidents,
    approval_diff1 = presidents) \%>\%
  DSD_Memory()
plot(stream, dim = 1, n = 120, method = "ts")

## apply a moving average filter to dimension 1
filteredStream <- stream \%>\%
  DSF_Convolve(kernel = filter_MA(5), dim = 2) \%>\%
  DSF_Convolve(kernel = filter_diff(1), dim = 3, na.rm = FALSE)
filteredStream

## resetting the filtered stream also resets the original stream
reset_stream(filteredStream)
get_points(filteredStream, n = 10)

reset_stream(filteredStream)
plot(filteredStream, dim = 1, n = 120, method = "ts")

## the second dimension is filtered
reset_stream(filteredStream)
plot(filteredStream, dim = 2, n = 120, method = "ts")

## look at different filters
plot(filter_MA(20), type = "s")
plot(filter_hamming(20), type = "s")
plot(filter_sinc(10, 100, width = 20), type = "s")
plot(filter_diff(1), type = "s")

}
\seealso{
\link[stats:filter]{stats::filter}

Other DSF: 
\code{\link{DSF_Downsample}()},
\code{\link{DSF_Func}()},
\code{\link{DSF_Scale}()},
\code{\link{DSF_dplyr}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
