% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Downsample.R
\name{DSF_Downsample}
\alias{DSF_Downsample}
\title{Downsample a Data Stream}
\usage{
DSF_Downsample(dsd, factor = 1)
}
\arguments{
\item{dsd}{The input stream as an \link{DSD} object.}

\item{factor}{the downsampling factor.}
}
\value{
An object of class \code{DSF_Downsample} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Creates a new stream that reduces the frequency of a given stream by a given factor.
}
\examples{
# memorize a stream
stream <- DSD_Memory(DSD_Gaussians(k = 3, noise = 0.05), n = 10)
get_points(stream, 10)

reset_stream(stream)

# create a filter that downsamples by a factor of 2 (returns points 1, 3, 5, ...)
downsampledStream <- DSF_Downsample(stream, factor = 2)

get_points(downsampledStream, 5)
}
\seealso{
Other DSF: 
\code{\link{DSF_Filter}()},
\code{\link{DSF_Func}()},
\code{\link{DSF_Scale}()},
\code{\link{DSF_dplyr}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
