% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_dplyr.R
\name{DSF_dplyr}
\alias{DSF_dplyr}
\title{Apply a dplyr Transformation to a Data Stream}
\usage{
DSF_dplyr(dsd, call = NULL)
}
\arguments{
\item{dsd}{A object of class \link{DSD}.}

\item{call}{a dplyr expression.}
}
\value{
An object of class \code{DSF_dplyr} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Applies a ...
}
\examples{
stream <- DSD_Gaussians(k = 3, d = 3)

stream2 <- stream \%>\%
  DSF_dplyr(select(X1, X2)) \%>\%
  DSF_dplyr(mutate(Xnew = X1 + X2)) \%>\%
  DSF_dplyr(filter(X1 > .5))
stream2

get_points(stream2, n = 10)
## Note: you get fewer points because of the filter operation.

}
\seealso{
\link[stats:filter]{stats::filter}

Other DSF: 
\code{\link{DSF_Downsample}()},
\code{\link{DSF_Filter}()},
\code{\link{DSF_Scale}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
