% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Func.R
\name{DSF_Func}
\alias{DSF_Func}
\title{Apply a Function to Transformation to a Data Stream}
\usage{
DSF_Func(dsd, func = NULL)
}
\arguments{
\item{dsd}{A object of class \link{DSD}.}

\item{func}{a function that takes a data.frame as the first argument and returns the transformed data.frame.}
}
\value{
An object of class \code{DSF_Func} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Applies an R function to transform to a data stream.
}
\details{
The function's first argument needs to be a data.frame representing points of the
data stream. The function will be called as \code{ps \%>\% your_function()}, where \code{ps} is the
data.frame with some points obtained using \code{\link[=get_points]{get_points()}} on the data stream source.
}
\examples{
stream <- DSD_Gaussians(k = 3, d = 3)

## Example 1: rename the columns
rename <- function(x, names) {
  colnames(x) <-  names
  x
}

stream2 <- stream \%>\% DSF_Func(rename(c("A", "B", "C")))
stream2

get_points(stream2, n = 5)

## Example 2: add a sum columns
sum_column <- function(x) {
  x$sum = rowSums(x)
 x
}

stream3 <- stream2 \%>\% DSF_Func(sum_column())
stream3
get_points(stream3, n = 5)
}
\seealso{
Other DSF: 
\code{\link{DSF_Downsample}()},
\code{\link{DSF_Filter}()},
\code{\link{DSF_Scale}()},
\code{\link{DSF_dplyr}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
