% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSF_Scale.R
\name{DSF_Scale}
\alias{DSF_Scale}
\alias{DSD_ScaleStream}
\title{Scale a Data Stream}
\usage{
DSF_Scale(
  dsd,
  dim = NULL,
  center = TRUE,
  scale = TRUE,
  n = 1000,
  reset = FALSE
)

DSD_ScaleStream(
  dsd,
  dim = NULL,
  center = TRUE,
  scale = TRUE,
  n = 1000,
  reset = FALSE
)
}
\arguments{
\item{dsd}{A object of class \link{DSD} that will be scaled.}

\item{dim}{integer vector; what dimensions should be scaled? Default is all.}

\item{center, scale}{logical or a numeric vector of length equal to the
number of columns used for centering/scaling (see function \link{scale}).}

\item{n}{The number of points used by \code{scale_stream()} to creating the centering/scaling}

\item{reset}{Try to reset the stream to its beginning after taking \code{n}
points for scaling.}
}
\value{
An object of class \code{DSF_Scale} (subclass of \link{DSF} and \link{DSD}).
}
\description{
Make an unscaled data stream into a scaled data stream.
}
\details{
\code{scale_stream()} estimates the values for centering and scaling (see
\link{scale} in \pkg{base}) using \code{n} points from the stream.
}
\section{Deprecated}{

\code{DSD_ScaleStream} is deprecated. Use \code{DSF_Scale} instead.
}

\examples{
## create a stream and apply scaling to z-scores using 100 points from the stream
stream <- DSD_Gaussians(k = 3, d = 2) \%>\% DSF_Scale(n = 100)
plot(stream)

## scale only the second dimension
stream <- DSD_Gaussians(k = 3, d = 2) \%>\% DSF_Scale(n = 100, dim = 2)
plot(stream)
}
\seealso{
\link{scale} in \pkg{base}

Other DSF: 
\code{\link{DSF_Convolve}()},
\code{\link{DSF_Downsample}()},
\code{\link{DSF_Func}()},
\code{\link{DSF_dplyr}()},
\code{\link{DSF}()}
}
\author{
Michael Hahsler
}
\concept{DSF}
