\name{DSFP}
\alias{DSFP}
\title{Abstract Class for Frequent Pattern Mining Algorithms for Data Streams}
\description{Abstract class for frequent pattern mining alforithms for data 
streams. Currently, 
\pkg{stream} does not implement frequent pattern mining algorithms.
}

%\details{
%}

\seealso{
	\code{\link{DST}}
}
