% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSO.R
\name{DSO}
\alias{DSO}
\title{Data Stream Operator Base Classes}
\usage{
DSO(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Abstract base classes for all DSO (Data Stream Operator) classes.
}
\details{
The \code{DSO} class cannot be instantiated, but it serve as a base
class from which other DSO classes inherit.

Data stream operators use \link{update} to process new data from the
\link{DSD} stream. The result of the operator can be obtained via \link{get_points}
and \link{get_weights} (if available).
}
\examples{
DSO()
}
\seealso{
\code{\link{DSO_Window}},
\code{\link{DSO_Sample}}
}
\author{
Michael Hahsler
}
