% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSOutlier.R
\docType{class}
\name{DSOutlier}
\alias{DSOutlier}
\alias{clean_outliers}
\alias{clean_outliers.DSOutlier}
\alias{recheck_outlier}
\alias{recheck_outlier.DSOutlier}
\alias{get_outlier_positions}
\alias{get_outlier_positions.DSOutlier}
\alias{noutliers}
\alias{noutliers.DSOutlier}
\title{Abstract Class for Outlier Detection Clusterers}
\usage{
DSOutlier(...)

clean_outliers(x, ...)

\method{clean_outliers}{DSOutlier}(x, ...)

recheck_outlier(x, outlier_correlated_id, ...)

\method{recheck_outlier}{DSOutlier}(x, outlier_correlated_id, ...)

get_outlier_positions(x, ...)

\method{get_outlier_positions}{DSOutlier}(x, ...)

noutliers(x, ...)

\method{noutliers}{DSOutlier}(x, ...)
}
\arguments{
\item{...}{further arguments.}

\item{x}{The DSC object.}

\item{outlier_correlated_id}{ids of outliers.}
}
\description{
The abstract class for all outlier detection clusterers. Cannot be
instantiated. An implementation is available in package
\pkg{streamMOA}.
}
\details{
\code{\link[=plot]{plot()}} has an extra logical argument to specify if outliers should be plotted
as red crosses.
}
\section{Functions}{
\itemize{
\item \code{clean_outliers}: forget detected outliers from the outlier detector.

\item \code{recheck_outlier}: Re-checks the outlier having \code{outlier_correlated_id}.
If this object is still an outlier, the method returns \code{TRUE}.

\item \code{get_outlier_positions}: Returns spatial positions of all current outliers.

\item \code{noutliers}: Returns the current number of outliers.
}}

\examples{
DSOutlier()
}
\seealso{
Other DST: 
\code{\link{DSAggregate}()},
\code{\link{DSClassify}()},
\code{\link{DSC}()},
\code{\link{DST}()},
\code{\link{predict}()},
\code{\link{update}()}
}
\author{
Dalibor Krleža
}
\concept{DST}
