\name{DST}
\alias{DST}
\title{Abstract Base Class for All Data Stream Mining Tasks}
\description{Abstract base class for 
all data stream mining tasks. Current tasks are
data stream clustering \code{DSC}, classification on data streams
\code{DSClass} and frequent pattern mining on data streams \code{DSFP}.
}
%\details{
%}

\seealso{
	\code{\link{DSC}},
	\code{\link{DSClass}},
	\code{\link{DSFP}}
}
