% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DST.R, R/description.R
\name{DST}
\alias{DST}
\alias{update}
\alias{update.DST}
\alias{description}
\alias{description.DSC}
\alias{description.DSD}
\alias{description.DSO}
\title{Conceptual Base Class for All Data Stream Mining Tasks}
\usage{
DST(...)

\method{update}{DST}(object, dsd, n = 1, ...)

description(x, ...)

\method{description}{DSC}(x, ...)

\method{description}{DSD}(x, ...)

\method{description}{DSO}(x, ...)
}
\arguments{
\item{...}{Further arguments.}

\item{dsd}{a data stream.}

\item{n}{the number of data points used for the update.}

\item{x, object}{an object of a concrete implementation of a DST.}
}
\description{
Conceptual base class for all data stream mining tasks. Current tasks are data
stream clustering \link{DSC}, outlier detection \link{DSOutlier}, classification on data streams
\link{DSClassify} and frequent pattern mining on data streams \link{DSFP}.
}
\details{
The method \code{update()} is used to update the data stream task with new datapoints coming from a data stream.
}
\section{Functions}{
\itemize{
\item \code{description}: Get a description of the task.
}}

\examples{
DST()
}
\seealso{
\link{DSC}, \link{DSOutlier}, \link{DSClassify}, \link{DSFP}, \link{DSO}
}
\author{
Michael Hahsler
}
