% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DST.R
\name{DST}
\alias{DST}
\title{Conceptual Base Class for All Data Stream Mining Tasks}
\usage{
DST(...)
}
\arguments{
\item{...}{Further arguments.}
}
\description{
Conceptual base class for all data stream mining tasks. Current tasks are data
stream clustering \link{DSC}, outlier detection \link{DSOutlier}, classification on data streams
\link{DSClassify} and frequent pattern mining on data streams \link{DSFP}.
}
\examples{
DST()
}
\author{
Michael Hahsler
}
