% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.R
\name{get_associated_prot}
\alias{get_associated_prot}
\title{Given a list of TFs, output associated proteins from BioGrid interacting with these TFs}
\usage{
get_associated_prot(key.TFs = NULL, db = "BioGrid", org = "human")
}
\arguments{
\item{key.TFs}{The list of intersted TF,s NULL by default}

\item{db}{The database of protein relations, including BioGrid, IntAct, STRING, and CORUM,
BioGrid by default}

\item{org}{The organism, human by default}
}
\description{
Given a list of TFs, output associated proteins from BioGrid interacting with these TFs
}
\keyword{internal}
