% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD.R
\name{get_points}
\alias{get_points}
\alias{remove_info}
\title{Get Points from a Data Stream Generator}
\usage{
get_points(x, n = 1L, outofpoints = "stop", info = FALSE, ...)

remove_info(points)
}
\arguments{
\item{x}{A \link{DSD} object.}

\item{n}{integer; request up to $n$ points from the stream.}

\item{outofpoints}{Action taken if less than \code{n} data points are
available. The default is to stop with an error.  Other actions that
might be supported by different \code{DSD} are:
\itemize{
\item \code{warn}: return the available points (maybe an empty data.frame) with a warning.
\item \code{ignore}: silently return the available points.
\item \code{block}: wait till enough data points are available.
}}

\item{info}{return additional columns with information about the data point (e.g., a known cluster assignment).}

\item{...}{Additional parameters to pass to the \code{get_points()} implementations.}

\item{points}{a data.frame with points.}
}
\value{
Returns a \link{data.frame} with (up to) \code{n} rows and as many columns as \code{x} produces.
}
\description{
Gets points from a \link{DSD} object.
}
\details{
Each DSD object has a unique way for creating/returning data points, but they all are
called through the generic function, \code{get_points()}. This is done by
using the S3 class system. See the man page for the specific \link{DSD} class on
the semantics for each implementation of \code{get_points()}.

Additional point information (e.g., known cluster/class assignment, noise status) can be requested
with \code{info = TRUE}. This information is returned as additional columns. The column names start with
\code{.} and are ignored by \link{DST} implementations. \code{remove_info()} is a convenience function to remove the
information columns.
}
\examples{
stream <- DSD_Gaussians()
points <- get_points(stream, n = 5, info = TRUE)
points

remove_info(points)
}
\seealso{
Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_NULL}()},
\code{\link{DSD_ReadCSV}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{DSF}()},
\code{\link{animate_data}()},
\code{\link{plot.DSD}()},
\code{\link{reset_stream}()},
\code{\link{write_stream}()}
}
\author{
Michael Hahsler
}
\concept{DSD}
