% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_stream-package.R
\docType{package}
\name{stream-package}
\alias{stream-package}
\title{stream: Infrastructure for Data Stream Mining}
\description{
A framework for data stream modeling and associated data mining tasks such as clustering and classification. The development of this package was supported in part by NSF IIS-0948893 and NIH R21HG005912. Hahsler et al (2017) \url{doi:10.18637/jss.v076.i14}.
}
\author{
Michael Hahsler
}
