% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_stream-package.R
\docType{package}
\name{stream-package}
\alias{stream-package}
\title{stream: Infrastructure for Data Stream Mining}
\description{
A framework for data stream modeling and associated data mining tasks such as clustering and classification.
}
\author{
Michael Hahsler
}
