% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stream.R
\name{write_stream}
\alias{write_stream}
\alias{write_stream.DSD}
\title{Write a Data Stream to a File}
\usage{
write_stream(
  dsd,
  file,
  n = 100,
  block = 100000L,
  class = FALSE,
  append = FALSE,
  sep = ",",
  header = FALSE,
  row.names = FALSE,
  write_outliers = FALSE,
  ...
)

\method{write_stream}{DSD}(
  dsd,
  file,
  n = 100,
  block = 100000L,
  class = FALSE,
  append = FALSE,
  sep = ",",
  header = FALSE,
  row.names = FALSE,
  write_outliers = FALSE,
  ...
)
}
\arguments{
\item{dsd}{The DSD object that will generate the data points for output.}

\item{file}{A file name or a R connection to be written to.}

\item{n}{The number of data points to be written.}

\item{block}{Write stream in blocks to improve file I/O speed.}

\item{class}{Save the class/cluster labels of the points as the last column.}

\item{append}{Append the data to an existing file.}

\item{sep}{The character that will separate attributes in a data point.}

\item{header}{A flag that determines if column names will be output
(equivalent to \code{col.names} in \code{write.table()}).}

\item{row.names}{A flag that determines if row names will be output.}

\item{write_outliers}{A flag that determines if outliers will be output.}

\item{...}{Additional parameters that are passed to \code{write.table()}.}
}
\value{
There is no value returned from this operation.
}
\description{
Writes points from a data stream DSD object to a file or a connection.
}
\examples{

# creating data and writing it to disk
stream <- DSD_Gaussians(k=3, d=5, outliers=1,
  outlier_options=list(outlier_horizon=10))
write_stream(stream, file="data.txt", n=10, class=TRUE, write_outliers=TRUE)

#file.show("data.txt")

# clean up
file.remove("data.txt")

}
\seealso{
\code{\link{write.table}}

Other DSD: 
\code{\link{DSD_BarsAndGaussians}()},
\code{\link{DSD_Benchmark}()},
\code{\link{DSD_Cubes}()},
\code{\link{DSD_Gaussians}()},
\code{\link{DSD_MG}()},
\code{\link{DSD_Memory}()},
\code{\link{DSD_ReadCSV}()},
\code{\link{DSD_ReadDB}()},
\code{\link{DSD_Target}()},
\code{\link{DSD_UniformNoise}()},
\code{\link{DSD_mlbenchData}()},
\code{\link{DSD_mlbenchGenerator}()},
\code{\link{DSD}()},
\code{\link{animate}},
\code{\link{get_points}()},
\code{\link{plot}()},
\code{\link{reset_stream}()}
}
\author{
Michael Hahsler, Dalibor Krleža
}
\concept{DSD}
